requirements.txt
flask==0.12
GitPython==2.1.3
